Added unit testing system for resource mngmnt server.
Added a unit testing framework to the server side of the
resource management tree. There is a base unit test class, then
child classes for vm-util, auth-util, db-util, and server-util that
contain groups of unit tests for each of those topics. There is a
master test program called regression.py that can be called with
different arguments to run any or all tests in this testing group.

We should consider using this model globally for our python testing.