Got a live devops API version working.
Completely redid the api documention for the devops request handler,
those updates are in the wiki and here locally. I git
devops_request_handler into a working state work 2 of the API endpoints,
GET /user/$username and  POST /user/. Both are tested but not fully. It
shouldn't be too hard to implement the rest and get this server up and
running.