Instance creation via API is now working, along with other calls.
Users can be created and listed with the API, credentials can be
validted and tokens grabbed, and instances can be created a listed. The
ssh-keys for the user that creates the instance are automatically added
to it so they can log-in right away without needing to change the root
password. Need to get snapshots and sshkeys implemented then the client
stuff and it'll be done.