API v.1 is implemented, tested with Curl. Client-side coming tomorrow.
The API is complete for version 0.1. Users can be added and updated,
sshkeys can be added to the DB and auto-added to new instances to allow
password free login for VMs. Virtual machines can be created from a raw
ubuntu image, or from an existing custom snapshot. Snapshots can be made
from existing instances. Token validation system is working well, and
all of the resources are available through GET requests.

All that's needed now is a simple init script to grab a users SSH key
and store it in the DB, and a few functions to wrap around the API.

3600+ handwritten LOC added in the last week.