Halfway done with the client script to handle all resources.
It turned out to kind of be a beast to write, but it's coming along
well. Right now I can use it to easily add users and grab info on them.
The setup functionality including the ssh-keygen process is tested and
working correctly, as are the user password and email update functions.
Right now you can grab info on current running instances, individually,
all at once, or all for a given user. Tomorrow I'll finish up the rest
of the calls and sync the branch with master.