API v1 client script is working and ready.
I just need to transfer the server-request-handler script from my
machine to a VM and switch the nameservers around to point
devops.praxyk.com at it then everyone will be set. I have tested the
functionality and it's pretty easy to make VMs and snapshots. I'll be
posting a tutorial with the script later.