I'm able to add and query users from the DB through the API.
I can run the api script locally using the --local flag and then add and
query user's from the user's table in the database. That means the
user's database class definition is working and the user's api-route is
working.. Next the transactions db-class and api-route class need to
worked out then the results.